requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.20.0
pillow==10.1.0
numpy>=1.26.0
opencv-python==4.8.1.78
scikit-image==0.24.0
scipy==1.11.4
python-multipart==0.0.6
python-dotenv>=1.0.0
torchxrayvision
torch
torchvision
google-generativeai>=0.8.0
